requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
streamlit-cropper
scikit-image
matplotlib
seaborn
scikit-learn
scipy
stqdm
prettytable
pandas